Removed redundant paramters. Separated source,sink and mapper. Removed dead code
