+++ icu_pipeline/mapper/source/__init__.py
@@ -10,7 +10,6 @@ from psycopg import sql
 
 from icu_pipeline.mapper.schema.fhir import AbstractFHIRSinkSchema
 from icu_pipeline.mapper.schema.ohdsi import AbstractOHDSISinkSchema
-from icu_pipeline.mapper.sink import AbstractSinkMapper, MappingFormat
 
 F = TypeVar("F", bound=AbstractFHIRSinkSchema)
 O = TypeVar("O", bound=AbstractOHDSISinkSchema)
@@ -35,42 +34,23 @@ class AbstractSourceMapper(ABC, Generic[F, O]):
         concept_id: str,
         concept_type: str,
         fhir_schema: Type[AbstractFHIRSinkSchema],
-        ohdsi_schema: Type[AbstractOHDSISinkSchema],
-        source_mapper_config: SourceMapperConfiguration,
-        sink_mapper: AbstractSinkMapper,
-        mapping_format: MappingFormat = MappingFormat.FHIR,
+        source_config: SourceMapperConfiguration,
     ) -> None:
         super().__init__()
 
         self._concept_id = concept_id
         self._concept_type = concept_type
         self._fhir_schema = fhir_schema
-        self._ohdsi_schema = ohdsi_schema
-        self._source_config = source_mapper_config
-        self._sink_mapper = sink_mapper
-        self._mapping_format = mapping_format
+        self._source_config = source_config
 
     def map(self) -> None:
-        # FHIR is default and later transformed if necessary
+        # FHIR is default and later transformed if necessary in a separate module
         return self.to_fihr()
-        # match self._mapping_format:
-        #     case MappingFormat.FHIR:
-        #         return self.to_fihr()
-        #     case MappingFormat.OHDSI:
-        #         return self.to_ohdsi()
-        #     case _:
-        #         raise KeyError(f"Couldn't match MappingFormat '{self._mapping_format}'")
 
     def to_fihr(self):
         for df in self.get_data():
             df = self._to_fihr(df).pipe(self._fhir_schema)
             yield df
-
-    def to_ohdsi(self):
-        for df in self.get_data():
-            df = self._to_ohdsi(df).pipe(self._ohdsi_schema)
-            self._sink_mapper.to_output_format(
-                df, self._ohdsi_schema, self._concept_id)
 
     @abstractmethod
     def get_data(self) -> Generator[pd.DataFrame, None, None]:
@@ -78,10 +58,6 @@ class AbstractSourceMapper(ABC, Generic[F, O]):
 
     @abstractmethod
     def _to_fihr(self, df: DataFrame) -> DataFrame[F]:
-        raise NotImplementedError
-
-    @abstractmethod
-    def _to_ohdsi(self, df: DataFrame) -> DataFrame[O]:
         raise NotImplementedError
